requirements.txt
unidecode
unicodecsv
# Excel support
xlrd
pycrypto
pysftp
wand
